requirements.txt
altair==5.5.0
annotated-types==0.7.0
appnope==0.1.4
asttokens==3.0.0
attrs==25.3.0
blinker==1.9.0
blis==1.3.0
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
click-default-group==1.2.4
cloudpathlib==0.21.0
comm==0.2.2
confection==0.1.5
cymem==2.0.11
debugpy==1.8.14
decorator==5.2.1
elastic-transport==8.17.1
elasticsearch==9.0.1
exceptiongroup==1.3.0
executing==2.2.0
faiss-cpu==1.11.0
filelock==3.18.0
fsspec==2025.3.2
gitdb==4.0.12
GitPython==3.1.44
hf-xet==1.1.1
huggingface-hub==0.31.1
idna==3.10
ipykernel==6.29.5
ipython==8.36.0
isodate==0.7.2
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.7.2
langcodes==3.5.0
language_data==1.3.0
lightgbm==4.6.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
murmurhash==1.0.12
narwhals==1.39.0
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.2.5
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.8
plotly==6.0.1
pluggy==1.5.0
preshed==3.0.9
prompt_toolkit==3.0.51
protobuf==6.30.2
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==20.0.0
pydantic==2.11.4
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.4.0
rdflib==7.1.4
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==14.0.0
rpds-py==0.24.0
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
sentence-transformers==4.1.0
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
smmap==5.0.2
spacy==3.8.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlite-fts4==1.0.3
sqlite-utils==3.38
srsly==2.5.1
stack-data==0.6.3
streamlit==1.45.1
streamlit-image-zoom==0.0.4
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
thinc==8.3.6
threadpoolctl==3.6.0
tokenizers==0.21.1
toml==0.10.2
torch==2.7.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.51.3
typer==0.15.3
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
wasabi==1.1.3
watchdog==6.0.0
wcwidth==0.2.13
weasel==0.4.1
Whoosh==2.7.4
wrapt==1.17.2
